Move admin related forms to a separate app
