+++ peacecorps/contenteditor/admin.py
@@ -1,0 +1,15 @@
+from django.contrib import admin
+from django.contrib.auth.admin import UserAdmin
+from django.contrib.auth.models import User
+
+from .form import (
+    StrictAdminPasswordChangeForm, StrictUserCreationForm)
+
+
+# Register your models here.
+class StrictUserAdmin(UserAdmin):
+    add_form = StrictUserCreationForm
+    change_password_form = StrictAdminPasswordChangeForm
+
+admin.site.unregister(User)
+admin.site.register(User, StrictUserAdmin)
